remove use of deprecated SubfieldBase
SubfieldBase was deprecated in Django 1.8 and was replaced by Field.from_db_value. This changes the behaviour of the Version and SpecField. When using SubfieldBase Field.to_python was called on assignment of the field. This is not the case anymore when using Field.from_db_value. See the Django documentation https://docs.djangoproject.com/en/1.9/howto/custom-model-fields/#changing-a-custom-field-s-base-class
This is also the reason for the full_clean calls in the tests.